Configuration changes made thread safe
