+++ client/src/main/java/com/orientechnologies/orient/client/remote/OStorageRemoteConfiguration.java
@@ -20,11 +20,11 @@
 
 package com.orientechnologies.orient.client.remote;
 
-import com.orientechnologies.orient.core.config.OStorageConfiguration;
+import com.orientechnologies.orient.core.config.OStorageConfigurationImpl;
 
 import java.nio.charset.Charset;
 
-public class OStorageRemoteConfiguration extends OStorageConfiguration {
+public class OStorageRemoteConfiguration extends OStorageConfigurationImpl {
 
   private static final long serialVersionUID = -3850696054909943272L;
   private String networkRecordSerializer;
