+++ core/src/main/java/com/orientechnologies/orient/core/config/OStoragePaginatedClusterConfiguration.java
@@ -28,13 +28,13 @@ public class OStoragePaginatedClusterConfiguration implements OStorageClusterCon
   public static final float              DEFAULT_GROW_FACTOR      = (float) 1.2;
   public float                           recordOverflowGrowFactor = DEFAULT_GROW_FACTOR;
   public float                           recordGrowFactor         = DEFAULT_GROW_FACTOR;
-  public String                          compression;
-  public String                          encryption;
-  public String                          encryptionKey;
+  public           String                    compression;
+  public           String                    encryption;
+  public           String                    encryptionKey;
   public transient OStorageConfiguration root;
-  public int                             id;
-  public String                          name;
-  public String                          location;
+  public           int                       id;
+  public           String                    name;
+  public           String                    location;
   public boolean                         useWal                   = true;
   public String                          conflictStrategy;
   private STATUS                         status                   = STATUS.ONLINE;
