+++ core/src/test/java/com/orientechnologies/orient/core/record/impl/BinaryComparatorEqualsTest.java
